Moved mock tests into their own class, fixed defect in the schema hash map, and updated the readme.md
