+++ src/test/java/com/cleo/labs/json_schema/SchemaValidation.java
@@ -32,10 +32,8 @@ public class SchemaValidation {
                     .put("action", schema("action.schema"))
                     .put("collection", schema("collection.schema"))
                     .put("common",schema("common.schema"))
-                    .put("connectionFile",schema("connectionFile.schema"))
                     .put("event", schema("event.schema"))
                     .put("job", schema("job.schema"))
-                    .put("resourcefolder", schema("resourcefolder.schema"))
                     .put("transfer", schema("transfer.schema"))
                     .build()
